+++ demos/main/src/main/java/com/google/android/reexoplayer2/demo/PlayerActivity.java
@@ -34,7 +34,6 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.google.android.exoplayer2.demo.R;
 import com.google.android.reexoplayer2.C;
 import com.google.android.reexoplayer2.C.ContentType;
 import com.google.android.reexoplayer2.DefaultRenderersFactory;